requirements.txt
requests
pygame
python-curses
discord-rpc.py
